+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-0.0.3
+0.1.0
 =====
+
+- Change {HandBrake::Title#chapters} from an Array to a Hash. This is
+  consistent with {HandBrake::Titles} and obviates the need for index
+  to chapter number conversions. {HandBrake::Title#all_chapters} is
+  the equivalent of the old {HandBrake::Title#chapters} method.
 
 0.0.2
 =====